requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-WTF>=1.0.0
Flask-Login>=0.6.2
WTForms>=3.0.0
pandas>=1.5.0
scikit-learn>=1.1.0
numpy>=1.21.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
Flask-Caching>=2.0.0
redis>=4.6.0
Werkzeug>=2.2.0
